@@ -1,17 +1,17 @@
 apply plugin: 'com.android.application'
 
 // Dependency versions
-def googleConglomerate = '27.1.0'
+def googleConglomerate = '1.0.0'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 30
     defaultConfig {
         applicationId "com.example.xianweili.newsapp"
         minSdkVersion 19
-        targetSdkVersion 27
+        targetSdkVersion 30
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
 
     compileOptions {
@@ -29,20 +29,20 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:' + googleConglomerate
-    implementation 'com.android.support:cardview-v7:' + googleConglomerate
-    implementation 'com.android.support:recyclerview-v7:' + googleConglomerate
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
+    implementation 'androidx.appcompat:appcompat:' + googleConglomerate
+    implementation 'androidx.cardview:cardview:' + googleConglomerate
+    implementation 'androidx.recyclerview:recyclerview:' + googleConglomerate
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
 
-    implementation 'com.jakewharton:butterknife:8.8.1'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
+    implementation 'com.jakewharton:butterknife:10.2.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.0'
 
     //Architecture Components
-    implementation "android.arch.lifecycle:extensions:1.1.1"
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
 
     //Paging
-    implementation "android.arch.paging:runtime:1.0.1"
+    implementation 'androidx.paging:paging-runtime:2.0.0'
 
     //Networking
     implementation 'com.squareup.retrofit2:retrofit:2.4.0'
